🐛(R4.1.3.12): Change wrong /var/log/faillog to /var/run/faillock
I think this is a transition error from an old implementation.
In CIS Ubuntu 20.04 v1.1 it was still /var/log/faillog, but in
CIS Ubuntu 20.04 v2.0 it has changed to /var/run/faillock and
hence in CIS Ubuntu 22.04 v1.0 it is also /var/run/faillock.

For Debian it is the same, for Debian 10 v1.0 it is /var/log/faillog,
but for Debian 11 v1.0 it is /var/run/faillock.

Signed-off-by: Fabian Raab <fabian.raab@siemens.com>